Create Gameboard & Player classes including test
Gameboard class with methods including to check
valid placement of a ship, place the ship,
generate random ship placement, and attack ships
placed on the board.

Player class enables player (either AI or Human)
to make a move and attack given cells. For the
human player, this will be linked to DOm events.